requirements.txt
astroid==2.2.5
dj-database-url==0.5.0
Django==2.2
djangorestframework==3.9.2
django-rest-swagger==2.1.2
gunicorn==19.9.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
whitenoise==4.1.2
wrapt==1.11.1
psycopg2==2.8.1
